feat: create UserPlaylist app
