+++ api/meow_sync_app/admin.py
@@ -5,4 +5,3 @@ admin.site.register(OAuthConnection)
 admin.site.register(OAuthState)
 admin.site.register(PlaylistSynchronization)
 admin.site.register(SongIdTranslation)
-admin.site.register(UserPlaylist)
