+++ api/meow_sync_app/models.py
@@ -64,14 +64,4 @@ class SongIdTranslation(AbstractBaseModel):
             ('user', 'spotify_id'),
         ]
 
-class UserPlaylist(AbstractBaseModel):
-    user = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        on_delete=models.CASCADE,
-    )
-    playlist_id = models.CharField(max_length=255)
-    provider = models.CharField(max_length=255, choices=MUSIC_PROVIDERS)
-    title = models.CharField(max_length=255)
-    description = models.TextField(blank=True, null=True)
-    author = models.CharField(max_length=255)
-    img_url = models.CharField(max_length=255, blank=True, null=True)
+
