+++ api/user_playlist/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class UserPlaylistConfig(AppConfig):
+    name = 'user_playlist'
