+++ api/user_playlist/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import UserPlaylist
+
+
+class UserPlaylistSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = UserPlaylist
+        fields = '__all__'
